Bound RelayChainStateCommitments by introducing a new bounded CurrentRelayChainStateRoots map and gradually draining the legacy unbounded map at 1000 entries per block.
