+++ modules/ismp/clients/parachain/client/Cargo.toml
@@ -38,11 +38,9 @@ features = [
     "sp-consensus-aura",
     "cumulus-pallet-parachain-system",
     "cumulus-primitives-core",
-    "pallet-migrations",
 ]
 
 [dev-dependencies]
-anyhow = { workspace = true }
 ismp-testsuite = { workspace = true, default-features = true }
 
 [dev-dependencies.polkadot-sdk]
@@ -53,7 +51,6 @@ features = [
     "frame-system",
     "pallet-balances",
     "pallet-timestamp",
-    "pallet-migrations",
     "staging-parachain-info",
     "sp-io",
     "sp-runtime",
