+++ parachain/runtimes/gargantua/Cargo.toml
@@ -114,7 +114,6 @@ features = [
     "staging-parachain-info",
     "parachains-common",
     "pallet-vesting",
-    "pallet-migrations",
     "sp-weights"
 ]
