+++ parachain/runtimes/nexus/Cargo.toml
@@ -129,7 +129,6 @@ features = [
     "staging-parachain-info",
     "parachains-common",
     "pallet-vesting",
-    "pallet-migrations",
 ]
 
 [features]
