+++ parachain/runtimes/nexus/src/lib.rs
@@ -401,7 +401,6 @@ impl frame_system::Config for Runtime {
 	/// The action to take on a Runtime Upgrade
 	type OnSetCode = cumulus_pallet_parachain_system::ParachainSetCode<Self>;
 	type MaxConsumers = frame_support::traits::ConstU32<16>;
-	type MultiBlockMigrator = Migrations;
 }
 
 impl pallet_timestamp::Config for Runtime {
@@ -766,24 +765,6 @@ impl pallet_vesting::Config for Runtime {
 	type BlockNumberProvider = System;
 }
 
-parameter_types! {
-	pub MbmServiceWeight: Weight = RuntimeBlockWeights::get().max_block.saturating_div(2);
-}
-
-impl pallet_migrations::Config for Runtime {
-	type RuntimeEvent = RuntimeEvent;
-	#[cfg(not(feature = "runtime-benchmarks"))]
-	type Migrations = (ismp_parachain::migration::MigrationV2<Runtime>,);
-	#[cfg(feature = "runtime-benchmarks")]
-	type Migrations = pallet_migrations::mock_helpers::MockedMigrations;
-	type CursorMaxLen = ConstU32<65_536>;
-	type IdentifierMaxLen = ConstU32<256>;
-	type MigrationStatusHandler = ();
-	type FailedMigrationHandler = frame_support::migrations::FreezeChainOnFailedMigration;
-	type MaxServiceWeight = MbmServiceWeight;
-	type WeightInfo = pallet_migrations::weights::SubstrateWeight<Runtime>;
-}
-
 impl pallet_bridge_airdrop::Config for Runtime {
 	type Currency = Balances;
 	type BridgeDropOrigin = EnsureRoot<Self::AccountId>;
@@ -1068,8 +1049,6 @@ mod runtime {
 	pub type CollatorManager = pallet_collator_manager;
 	#[runtime::pallet_index(94)]
 	pub type IntentsCoprocessor = pallet_intents_coprocessor;
-	#[runtime::pallet_index(95)]
-	pub type Migrations = pallet_migrations;
 
 	// consensus clients
 	#[runtime::pallet_index(254)]
@@ -1115,7 +1094,6 @@ mod benches {
 		[pallet_preimage, Preimage]
 		[pallet_vesting, Vesting]
 		[pallet_token_gateway, TokenGateway]
-		[pallet_migrations, Migrations]
 	);
 }
