+++ parachain/simtests/src/lib.rs
@@ -3,5 +3,5 @@ mod intents_rpc;
 mod migration_test;
 mod pallet_ismp;
 mod pallet_mmr;
-mod relay_state_migration_test;
+mod relay_state_drain_test;
 mod token_allocation;
